+++ demo/demo_script_04_stallguard.py
@@ -8,8 +8,8 @@ test file for testing the StallGuard feature
 
 import time
 try:
-    from src.tmc2209.tmc_2209 import *
-    from src.tmc2209._tmc_gpio_board import Board
+    from src.tmc_driver.tmc_2209 import *
+    from src.tmc_driver._tmc_gpio_board import Board
 except ModuleNotFoundError:
     from TMC_2209.TMC_2209_StepperDriver import *
     from TMC_2209._TMC_2209_GPIO_board import Board
@@ -33,12 +33,12 @@ if BOARD == Board.NVIDIA_JETSON:
         Nvidia Jetson has nuances with the parameter pull_up_down for pin_stallguard:
         https://github.com/NVIDIA/jetson-gpio/issues/5''')
 if BOARD == Board.RASPBERRY_PI:
-    tmc = TMC_2209(21, 16, 20)
+    tmc = Tmc2209(21, 16, 20)
 elif BOARD == Board.RASPBERRY_PI5:
-    tmc = TMC_2209(21, 16, 20, serialport="/dev/ttyAMA0")
+    tmc = Tmc2209(21, 16, 20, serialport="/dev/ttyAMA0")
 else:
     # just in case
-    tmc = TMC_2209(21, 16, 20)
+    tmc = Tmc2209(21, 16, 20)
