+++ demo/demo_script_07_threads.py
@@ -8,8 +8,8 @@ test file for testing movement of motors with threads
 
 import time
 try:
-    from src.tmc2209.tmc_2209 import *
-    from src.tmc2209._tmc_gpio_board import Board
+    from src.tmc_driver.tmc_2209 import *
+    from src.tmc_driver._tmc_gpio_board import Board
 except ModuleNotFoundError:
     from TMC_2209.TMC_2209_StepperDriver import *
     from TMC_2209._TMC_2209_GPIO_board import Board
@@ -26,14 +26,14 @@ print("---")
 # use your pins for pin_en, pin_step, pin_dir here
 #-----------------------------------------------------------------------
 if BOARD == Board.RASPBERRY_PI:
-    tmc1 = TMC_2209(21, 16, 20)
+    tmc1 = Tmc2209(21, 16, 20)
 elif BOARD == Board.RASPBERRY_PI5:
-    tmc1 = TMC_2209(21, 16, 20, serialport="/dev/ttyAMA0")
+    tmc1 = Tmc2209(21, 16, 20, serialport="/dev/ttyAMA0")
 elif BOARD == Board.NVIDIA_JETSON:
-    tmc1 = TMC_2209(13, 6, 5, serialport="/dev/ttyTHS1")
+    tmc1 = Tmc2209(13, 6, 5, serialport="/dev/ttyTHS1")
 else:
     # just in case
-    tmc1 = TMC_2209(21, 16, 20)
+    tmc1 = Tmc2209(21, 16, 20)
 
 tmc_driverlist = [tmc1]
