+++ demo/demo_script_10_only_step_pin.py
@@ -8,8 +8,8 @@ test file for testing basic movement
 
 import time
 try:
-    from src.tmc2209.tmc_2209 import *
-    from src.tmc2209._tmc_gpio_board import Board
+    from src.tmc_driver.tmc_2209 import *
+    from src.tmc_driver._tmc_gpio_board import Board
 except ModuleNotFoundError:
     from TMC_2209.TMC_2209_StepperDriver import *
     from TMC_2209._TMC_2209_GPIO_board import Board
@@ -28,14 +28,14 @@ print("---")
 # use your pins for pin_en, pin_step, pin_dir here
 #-----------------------------------------------------------------------
 if BOARD == Board.RASPBERRY_PI:
-    tmc = TMC_2209(pin_step=16, loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(pin_step=16, loglevel=Loglevel.DEBUG)
 elif BOARD == Board.RASPBERRY_PI5:
-    tmc = TMC_2209(pin_step=16, serialport="/dev/ttyAMA0", loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(pin_step=16, serialport="/dev/ttyAMA0", loglevel=Loglevel.DEBUG)
 elif BOARD == Board.NVIDIA_JETSON:
-    tmc = TMC_2209(pin_step=13, serialport="/dev/ttyTHS1", loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(pin_step=13, serialport="/dev/ttyTHS1", loglevel=Loglevel.DEBUG)
 else:
     # just in case
-    tmc = TMC_2209(pin_step=21, loglevel=Loglevel.DEBUG)
+    tmc = Tmc2209(pin_step=21, loglevel=Loglevel.DEBUG)
