+++ src/tmc_driver/reg/bitfields/_tmc_2209_sgthrs.py
