+++ src/tmc_driver/reg/bitfields/_tmc_220x_ihold_irun.py
