+++ src/tmc_driver/tmc_2208.py
@@ -1,0 +1,9 @@
+"""TMC_2208 stepper driver module
+"""
+
+from .tmc_220x import Tmc220x
+
+
+
+class Tmc2208(Tmc220x):
+    """TMC_2209"""
